Extract fixture frontend/app.rb
