+++ spec/lib/msplex/resource/frontend_spec.rb
@@ -54,55 +54,7 @@ ELEMENTS
         subject { frontend.app_rb }
 
         it "should generate app.rb" do
-          expect(subject).to eq <<-APPRB
-class App < Sinatra::Base
-  configure do
-    use Rack::Session::Cookie, expire_after: 3600, secret: "salt"
-    use Rack::Csrf, raise: true
-    Slim::Engine.default_options[:pretty] = true
-  end
-
-  configure :development do
-    register Sinatra::Reloader
-  end
-
-  helpers do
-    def csrf_meta_tag
-      Rack::Csrf.csrf_metatag(env)
-    end
-
-    def param_str(parameters)
-      parameters.map { |key, value| key.to_s + "=" + CGI.escape(value.to_s) }.join("&")
-    end
-
-    def http_get(endpoint, parameters = {})
-      uri = URI.parse(parameters.length > 0 ? endpoint + "?" + param_str(parameters) : endpoint)
-      JSON.parse(Net::HTTP.get_response(uri).body, symbolize_names: true)
-    rescue
-      { error: true }
-    end
-
-    def http_post(endpoint, parameters)
-      uri = URI.parse(endpoint)
-      JSON.parse(Net::HTTP.post_form(uri, parameters).body, symbolize_names: true)
-    rescue
-      { error: true }
-    end
-
-    def endpoint_of(service, action)
-      "http://" << service << "/" << action
-    end
-  end
-
-  get "/" do
-    slim :index, locals: { users: http_get(endpoint_of("user", "users/list")) }
-  end
-
-  get "/search" do
-    slim :search, locals: {  }
-  end
-end
-APPRB
+          expect(subject).to eq open(fixture_path("frontend/app.rb")).read
         end
       end
